+++ pages/getFormMeta.php
@@ -56,14 +56,7 @@ if(empty($_GET['embed'])){
 
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <script type='text/javascript'>
-	if ((navigator.userAgent.match(/(phone|pad|pod|iPhone|iPod|ios|iPad|Android|Mobile|BlackBerry|IEMobile|MQQBrowser|JUC|Fennec|wOSBrowser|BrowserNG|WebOS|Symbian|Windows Phone)/i))){
-			if(screen.height>=1080)
-				document.write('<meta name="viewport" content="width=device-width, initial-scale=2.0, minimum-scale=1.0, maximum-scale=3.0, target-densityDpi=device-dpi, user-scalable=yes">');
-			else	
-				document.write('<meta name="viewport" content="width=device-width, initial-scale=0.5, minimum-scale=0.5, maximum-scale=3.0, target-densityDpi=device-dpi, user-scalable=yes">');
-	}
-  </script>
+
   <meta name="description" content="<?php echo $product.' - '.$ver;?>">
   <title><?php echo $info['name'];?></title>
   <link rel="icon" type="image/png" sizes="32x32" href="/img/favicon-32x32.png">
@@ -100,108 +93,138 @@ if(empty($_GET['embed'])){
 
 </style>
 
-
-
-
-<table class="table table-bordered"  cellspacing="0">
-  <tr>
-    <td colspan="2"><h1 class="mgmIngHeader mgmIngHeader-with-separator"><?=$info['name']?></h1><span class="mgmIngHeaderCAS"><?=$info['product_name']?></span></td>
-  </tr>
-</table>
+<div class="ml2">
+	<h1 class="mgmIngHeader mgmIngHeader-with-separator"><?=$info['name']?></h1><span class="mgmIngHeaderCAS"><?=$info['product_name']?></span>
+</div>
+<div class="dropdown-divider"></div>
 <?php } ?>
 
 <script src="/js/bootstrap-tagsinput.js"></script> 
 <link href="/css/bootstrap-tagsinput.css" rel="stylesheet" />
 
-<div id="set_msg"></div>
-<table class="table table-bordered" id="formula_metadata">
-  <tr>
-    <td>Formula Name:</td>
-    <td data-name="name" class="name" data-type="text" align="left" data-pk="<?php echo $info['id'];?>" width="80%"><?php echo $info['name'];?></td>
-  </tr>
-  <tr>
-    <td>Product Name:</td>
-    <td data-name="product_name" class="product_name" data-type="text" align="left" data-pk="<?php echo $info['id'];?>"><?php echo $info['product_name'];?></td>
-  </tr>
-  <tr>
-    <td><a href="#" rel="tip" title="When enabled, formula is protected against deletion. By enabling this, a formula revision will be automatically created.">Protected:</a></td>
-    <td><input name="isProtected" type="checkbox" id="isProtected" value="1" <?php if($info['isProtected'] == '1'){; ?> checked="checked"  <?php } ?>/></td>
-  </tr>
-  <tr>
-    <td>Customer:</td>
-    <td><select name="customer" id="customer" class="form-control ellipsis">
+<div class="form-horizontal card" id="formula_metadata">
+ <div id="set_msg"></div>
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="formula_name">Formula Name:</label>
+    <div class="col-sm-3">
+      <a href="#" data-name="name" class="name" id="formula_name" data-pk="<?php echo $info['id'];?>"><?php echo $info['name'];?></a>
+    </div>
+  </div>
+
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="product_name">Product Name:</label>
+    <div class="col-sm-3">
+     	<a href="#" data-name="product_name" class="product_name" id="product_name" data-pk="<?php echo $info['id'];?>"><?php echo $info['product_name'] ?: $info['name'];?></a>
+    </div>
+  </div>
+
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="isProtected">Protected:</label>
+    <div class="col-sm-3">
+     	<input name="isProtected" type="checkbox" id="isProtected" value="1" <?php if($info['isProtected'] == '1'){; ?> checked="checked"  <?php } ?>/>
+        <i class="fa-solid fa-circle-info" rel="tip" title="When enabled, formula is protected against deletion. By enabling this, a formula revision will be automatically created."></i>
+    </div>
+  </div>
+  
+   <div class="form-group">
+    <label class="control-label col-sm-3" for="customer">Customer:</label>
+    <div class="col-sm-3">
+     	<select name="customer" id="customer" class="form-control ellipsis">
       <option value="0">Internal use</option>
       <?php foreach ($customer as $c) {?>
       <option value="<?=$c['id'];?>" <?php echo ($info['customer_id']==$c['id'])?"selected=\"selected\"":""; ?>><?php echo $c['name'];?></option>
       <?php }	?>
-    </select></td>
-  </tr>
-  <tr>
-    <td>View:</td>
-    <td><select name="defView" id="defView" class="form-control">
+    </select>
+    </div>
+  </div>
+  
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="defView">Default view:</label>
+    <div class="col-sm-3">
+     	<select name="defView" id="defView" class="form-control">
 			  <option value="1" <?php if($info['defView']=="1") echo 'selected="selected"'; ?> >Ingredient Properties</option>
 			  <option value="2" <?php if($info['defView']=="2") echo 'selected="selected"'; ?> >Ingredient Notes</option>
-          </select></td>
-  </tr>
-  <tr>
-    <td>Profile:</td>
-    <td><a href="#" id="profile" data-type="select" data-pk="<?php echo $info['id'];?>" data-title="Select profile"></a></td>
-  </tr>
-  <tr>
-  <tr>
-    <td>Tags:</td>
-    <td> <input type="text" class="form-control col-xs-3 control-label" id="tagsinput" data-role="tagsinput" /></td>
-  </tr>
-  <tr>
-    <td>Purpose:</td>
-    <td>
-    <select name="catClass" id="catClass" class="form-control ellipsis">
-	<option></option>
-	<?php foreach ($cats as $IFRACategories) {?>
-	<option value="cat<?php echo $IFRACategories['name'];?>" <?php echo ($info['catClass']=='cat'.$IFRACategories['name'])?"selected=\"selected\"":""; ?>><?php echo 'Cat'.$IFRACategories['name'].' - '.$IFRACategories['description'];?></option>
-	<?php }	?>
-    </select></td>
-  </tr>
-  <tr>
-    <td>Final type:</td>
-    <td>
-    <select name="finalType" id="finalType" class="form-control ellipsis">  
+          </select>
+    </div>
+ </div>  
+  
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="profile">Profile:</label>
+    <div class="col-sm-3">
+		<a href="#" id="profile" data-type="select" data-pk="<?php echo $info['id'];?>" data-title="Select profile"></a>
+    </div>
+ </div>  
+  
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="tagsinput">Tags:</label>
+    <div class="col-sm-3">
+		<input type="text" class="form-control col-xs-3 control-label" id="tagsinput" data-role="tagsinput" />
+    </div>
+ </div>
+
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="catClass">Purpose:</label>
+    <div class="col-sm-3">
+		<select name="catClass" id="catClass" class="form-control ellipsis">
+            <option></option>
+            <?php foreach ($cats as $IFRACategories) {?>
+            <option value="cat<?php echo $IFRACategories['name'];?>" <?php echo ($info['catClass']=='cat'.$IFRACategories['name'])?"selected=\"selected\"":""; ?>><?php echo 'Cat'.$IFRACategories['name'].' - '.$IFRACategories['description'];?></option>
+            <?php }	?>
+        </select>
+    </div>
+ </div>
+  
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="finalType">Final type:</label>
+    <div class="col-sm-3">
+		<select name="finalType" id="finalType" class="form-control ellipsis">  
             <option value="100">Concentrated (100%)</option>
 	 		<?php foreach ($fTypes as $fType) {?>
-	<option value="<?php echo $fType['concentration'];?>" <?php echo ($info['finalType']==$fType['concentration'])?"selected=\"selected\"":""; ?>><?php echo $fType['name'].' ('.$fType['concentration'];?>%)</option>
-	<?php }	?>	
-    </select>
-    </td>
-  </tr>
-  <tr>
-    <td>Status:</td>
-    <td>
-    <select name="status" id="status" class="form-control ellipsis">  
-        <option value="0" <?php if($info['status'] == "0"){ echo 'selected';}?>>Scheduled</option>
-        <option value="1" <?php if($info['status'] == "1"){ echo 'selected';}?>>Under Developent</option>
-        <option value="2" <?php if($info['status'] == "2"){ echo 'selected';}?>>Under Evaluation</option>
-        <option value="3" <?php if($info['status'] == "3"){ echo 'selected';}?>>In Production</option>
-        <option value="4" <?php if($info['status'] == "4"){ echo 'selected';}?>>To be reformulated</option>
-        <option value="5" <?php if($info['status'] == "5"){ echo 'selected';}?>>Failure</option>
-    </select>    
-    </td>
-  </tr>
-  <tr>
-    <td>Gender:</td>
-    <td><a href="#" id="sex" data-type="select" data-pk="<?php echo $info['id'];?>" data-title="Select sex"></a></td>
-  </tr>
-  <tr>
-    <td>Picture:</td>
-    <td>
-      <input type="file" name="doc_file" id="doc_file" />
-      <input type="submit" name="button" class="btn btn-primary" id="pic_upload" value="Upload">
-    </td>
-  </tr>
-  <tr>
-    <td>Notes:</td>
-    <td data-name="notes" class="notes" data-type="textarea" align="left" data-pk="<?php echo $info['id'];?>"><?php echo $info['notes'];?></td>
-  </tr>
-</table>
+			<option value="<?php echo $fType['concentration'];?>" <?php echo ($info['finalType']==$fType['concentration'])?"selected=\"selected\"":""; ?>><?php echo $fType['name'].' ('.$fType['concentration'];?>%)</option>
+			<?php }	?>	
+    	</select>
+    </div>
+ </div>
+
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="finalType">Status:</label>
+    <div class="col-sm-3">
+        <select name="status" id="status" class="form-control ellipsis">  
+            <option value="0" <?php if($info['status'] == "0"){ echo 'selected';}?>>Scheduled</option>
+            <option value="1" <?php if($info['status'] == "1"){ echo 'selected';}?>>Under Developent</option>
+            <option value="2" <?php if($info['status'] == "2"){ echo 'selected';}?>>Under Evaluation</option>
+            <option value="3" <?php if($info['status'] == "3"){ echo 'selected';}?>>In Production</option>
+            <option value="4" <?php if($info['status'] == "4"){ echo 'selected';}?>>To be reformulated</option>
+            <option value="5" <?php if($info['status'] == "5"){ echo 'selected';}?>>Failure</option>
+        </select>
+    </div>
+ </div>
+
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="sex">Gender:</label>
+    <div class="col-sm-3">
+		<a href="#" id="sex" data-type="select" data-pk="<?php echo $info['id'];?>" data-title="Select gender"></a>
+    </div>
+ </div>
+ 
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="doc_file">Picture:</label>
+    <div class="col-sm-3">
+		<input type="file" name="doc_file" id="doc_file" />
+      	<input type="submit" name="button" class="btn btn-primary" id="pic_upload" value="Upload">
+    </div>
+ </div>
+ 
+ <div class="form-group">
+    <label class="control-label col-sm-3" for="notes">Notes:</label>
+    <div class="col-sm-3">
+		<a href="#" data-name="notes" class="notes" data-type="textarea" id="notes" data-pk="<?php echo $info['id'];?>"><?php echo $info['notes']?: 'None';?></a>
+    </div>
+ </div>
+ 
+  
+</div>
+
 
 <script type="text/javascript" language="javascript" >
 $(document).ready(function(){
@@ -211,39 +234,47 @@ $('[rel=tip]').tooltip({placement: 'auto'});
 
 $('#formula_metadata').editable({
   container: 'body',
-  selector: 'td.name',
+  selector: 'a.name',
   url: "/pages/update_data.php?action=rename&fid=<?=$info['fid']?>",
   title: 'Name',
-  type: "POST",
   mode: 'inline',
-  dataType: 'json',
-      success: function(response) {				
-	  	$('#set_msg').html(response);        
+  ajaxOptions: { 
+  	dataType: 'json'
+  },
+  validate: function(value){
+  	if($.trim(value) == ''){
+		return 'This field is required';
+   	}
+  },
+  success: function(response) {	
+  	if(response.success){
+		msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + response.success + '</strong></div>';
+	}else{
+		msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + response.error + '</strong></div>';
+	}
+	  	$('#set_msg').html(msg);        
     },
 
 });
   
 $('#formula_metadata').editable({
   container: 'body',
-  selector: 'td.notes',
+  selector: 'a.notes',
+  emptytext: 'None',
   url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
   title: 'Notes',
-  type: "POST",
-  mode: 'inline',
-  dataType: 'json',
-      success: function(response, newValue) {
-        if(response.status == 'error') return response.msg; 
-    },
+  mode: 'inline'
+
 
 });
   
 $('#formula_metadata').editable({
   container: 'body',
-  selector: 'td.product_name',
+  selector: 'a.product_name',
   url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
   title: 'Product Name',
-  type: "POST",
   mode: 'inline',
+  emptytext: 'None',
   dataType: 'json',
       success: function(response, newValue) {
         if(response.status == 'error') return response.msg; 
@@ -252,10 +283,10 @@ $('#formula_metadata').editable({
 });
 
 $('#profile').editable({
-value: "<?php echo $info['profile'];?>",
-title: 'Profile',
-url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
-source: [
+	value: "<?php echo $info['profile'];?>",
+	title: 'Profile',
+	url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
+	source: [
 		<?php foreach ($fcat as $cat) { if($cat['type'] == 'profile'){?>		
 		 {value: '<?=$cat['cname']?>', text: '<?=$cat['name']?>'},
 		<?php } }?>
@@ -263,11 +294,12 @@ source: [
 });
 
 $('#sex').editable({
-value: "<?php echo $info['sex'];?>",
-url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
-source: [
+	value: "<?php echo $info['sex'];?>",
+	emptytext: 'Please select',
+	url: "/pages/update_data.php?formulaMeta=<?=$info['fid']?>",
+	source: [
 		 <?php foreach ($fcat as $cat) { if($cat['type'] == 'sex'){?>		
-		 {value: '<?=$cat['cname']?>', text: '<?=$cat['name']?>'},
+		 	{value: '<?=$cat['cname']?>', text: '<?=$cat['name']?>'},
 		<?php } }?>
 	   ]
 });