+++ pages/top.php
@@ -56,7 +56,7 @@ if($pv_meta['schema_ver'] < $db_ver){
               </a>
               <div class="dropdown-menu dropdown-menu-right shadow animated--grow-in" aria-labelledby="userDropdown">
               
-				<a class="dropdown-item popup-link" href="pages/editUser.php">
+				<a class="dropdown-item popup-link" href="/pages/editUser.php">
                   <i class="fas fa-user fa-sm fa-fw mr-2 text-gray-400"></i>
                   Edit my details
                 </a>